+++ src/widgets/history/styles/book/content_container/Award.css
