+++ src/widgets/history/styles/book/content_container/Content.css
